requirements.txt
py-cord
py-cord[voice]
python-dotenv
aiodns
brotlipy
orjson
